# integration-layer-04-openapi-spec.yaml
openapi: 3.0.0
info:
  title: Oli Tools Universal Integration API
  description: |
    Complete API documentation for Oli Tools integration layer.
    
    Supports:
    - Zapier, Make.com, n8n, GoHighLevel (GHL)
    - Inbound webhooks (Zapier → Oli)
    - Outbound webhooks (Oli → external services)
    - OAuth authentication (connect external platforms)
    
    **Authentication:** All requests require Bearer token in Authorization header
    
    ```
    Authorization: Bearer {api_token}
    ```
  version: 1.0.0
  contact:
    name: Oli Support
    email: support@oli.tools
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.oli.tools
    description: Production API
  - url: https://staging-api.oli.tools
    description: Staging API
  - url: http://localhost:3000
    description: Local development

# ============================================================================
# AUTHENTICATION
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 access token or API token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - error
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastTriggeredAt:
          type: string
          format: date-time
...
paths:
  /api/webhooks/v1/{toolKey}/{action}:
    post:
      summary: Send webhook to Oli tool
      description: |
        Inbound webhook handler. Send data from Zapier, Make, n8n, or GHL to trigger actions in Oli tools.
        
        **Example:** Zapier receives a new Stripe payment → POST to OliCommerce `create_contact` action
      parameters:
        - in: path
          name: toolKey
          required: true
          schema:
            type: string
            enum: [oliops, olicommerce, oliflow, oliexplore, oli-locator, olisalestrack]
          description: Target Oli tool
        - in: path
          name: action
          required: true
          schema:
            type: string
            enum: [create_contact, create_lead, trigger_workflow, publish_post, record_sale, update_contact]
          description: Action to perform
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Action-specific payload
              example:
                email: user@example.com
                name: John Doe
                phone: "+1-555-0100"
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  requestId:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - Inbound Webhooks

# ============================================================================
# PATHS: OUTBOUND WEBHOOKS (Oli → External)
# ============================================================================
...
  /api/webhooks:
    get:
      summary: List webhooks
      description: Get all registered webhooks for a user
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: toolKey
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  count:
                    type: integer
      tags:
        - Outbound Webhooks
...
  /api/webhooks/logs:
    get:
      summary: Get webhook logs
      description: View recent webhook delivery logs (debugging)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [success, error]
        - in: query
          name: tool
          schema:
            type: string
      responses:
        '200':
          description: Webhook logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                  count:
                    type: integer
      tags:
        - Debugging
security:
  - BearerAuth: []
...